OnWork; add credit history input for users

+++ migrations/0006_RelacionPrestamoSeguro.sql
@@ -1,4 +1,4 @@
 CREATE TABLE RelacionPrestamoSeguro(
-    fkPrestamo INT NOT NULL REFERENCES Prestamo(ID),
-    fkSeguro INT NOT NULL REFERENCES Seguro(ID)
+    fkPrestamo BIGINT NOT NULL REFERENCES Prestamo(ID),
+    fkSeguro BIGINT NOT NULL REFERENCES Seguro(ID)
 )
